# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Cargo transfer to ISS",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/",
    "paths": {
...
        "/api/flights": {
            "get": {
                "description": "Возвращает все полеты с фильтрацией по статусу и дате формирования",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Полеты"
                ],
                "summary": "Получить все полеты",
                "parameters": [
                    {
                        "type": "string",
                        "description": "статус полета",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "начальная дата формирования",
                        "name": "form_date_start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "конечная дата формирвания",
                        "name": "form_date_end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.AllFlightsResponse"
                        }
                    }
                }
            }
        },
...
        "/api/flights/{flight_id}": {
            "get": {
                "description": "Возвращает подробную информацию о полете",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Полеты"
                ],
                "summary": "Получить один полет",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id полета",
                        "name": "flight_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.FlightResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Позволяет изменить тип ракеты",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Полеты"
                ],
                "summary": "Указать тип ракеты",
                "parameters": [
                    {
                        "description": "Тип ракеты",
                        "name": "rocket_type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.SwaggerUpdateFlightRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.UpdateFlightResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет полет по id",
                "tags": [
                    "Полеты"
                ],
                "summary": "Удалить полет",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/flights/{flight_id}/change_cargo/{cargo_id}": {
            "put": {
                "description": "Изменить количество груза в полете",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Полеты"
                ],
                "summary": "Изменить количество груза в полете",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id полета",
                        "name": "flight_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id груза",
                        "name": "cargo_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "количество груза",
                        "name": "quantity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/flights/{flight_id}/delete_cargo/{cargo_id}": {
            "delete": {
                "description": "Удалить груз из полета",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Полеты"
                ],
                "summary": "Удалить груз из полета",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id груза",
                        "name": "cargo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.AllCargosResponse"
                        }
                    }
                }
            }
        },
        "/api/flights/{flight_id}/moderator_confirm": {
            "put": {
                "description": "Завершить или отклонить полет модератором",
                "tags": [
                    "Полеты"
                ],
                "summary": "Завершить полет",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id полета",
                        "name": "flight_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "подтвердить",
                        "name": "confirm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.FlightOutput"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.SwaggerUpdateFlightRequest": {
            "type": "object",
            "properties": {
                "rocket_type": {
                    "type": "string"
                }
            }
        },
        "ds.Cargo": {
            "type": "object",
            "required": [
                "capacity",
                "category",
                "description",
                "en_name",
                "name",
                "price",
                "weight"
            ],
            "properties": {
                "capacity": {
                    "description": "m^3",
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "en_name": {
                    "type": "string"
                },
                "isDeleted": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                },
                "price": {
                    "description": "Rubles",
                    "type": "integer"
                },
                "uuid": {
                    "type": "string"
                },
                "weight": {
                    "description": "kg",
                    "type": "number"
                }
            }
        },
...
        "schemes.AllCargosResponse": {
            "type": "object",
            "properties": {
                "cargos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Cargo"
                    }
                }
            }
        },
        "schemes.AllFlightsResponse": {
            "type": "object",
            "properties": {
                "flights": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemes.FlightOutput"
                    }
                }
            }
        },
        "schemes.FlightOutput": {
            "type": "object",
            "properties": {
                "completion_date": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "customer": {
                    "type": "string"
                },
                "formation_date": {
                    "type": "string"
                },
                "moderator": {
                    "type": "string"
                },
                "rocket_type": {
                    "type": "string"
                },
                "shipment_status": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "schemes.FlightResponse": {
            "type": "object",
            "properties": {
                "cargos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Cargo"
                    }
                },
                "flight": {
                    "$ref": "#/definitions/schemes.FlightOutput"
                }
            }
        },
...
        "schemes.UpdateFlightResponse": {
            "type": "object",
            "properties": {
                "flight": {
                    "$ref": "#/definitions/schemes.FlightOutput"
                }
            }
        }
    }
}